requirements/production.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    ./bin/compile_dependencies.sh
#
alabaster==0.7.12
    # via
    #   -r requirements/base.txt
    #   sphinx
asgiref==3.5.2
    # via
    #   -r requirements/base.txt
    #   django
async-timeout==4.0.2
    # via
    #   -r requirements/base.txt
    #   redis
attrs==22.1.0
    # via
    #   -r requirements/base.txt
    #   jsonschema
babel==2.10.3
    # via
    #   -r requirements/base.txt
    #   sphinx
certifi==2022.9.24
    # via
    #   -r requirements/base.txt
    #   django-simple-certmanager
    #   requests
cffi==1.15.1
    # via
    #   -r requirements/base.txt
    #   cryptography
charset-normalizer==2.1.1
    # via
    #   -r requirements/base.txt
    #   requests
cryptography==38.0.1
    # via
    #   -r requirements/base.txt
    #   cryptograpy
    #   pyopenssl
cryptograpy==0.0.0
    # via
    #   -r requirements/base.txt
    #   django-simple-certmanager
deprecated==1.2.13
    # via
    #   -r requirements/base.txt
    #   redis
django==3.2.15
    # via
    #   -r requirements/base.txt
    #   django-axes
    #   django-capture-on-commit-callbacks
    #   django-choices
    #   django-filter
    #   django-markup
    #   django-privates
    #   django-redis
    #   django-relativedelta
    #   django-rest-framework-condition
    #   django-sendfile2
    #   django-simple-certmanager
    #   django-solo
    #   djangorestframework
    #   drf-nested-routers
    #   drf-spectacular
    #   notifications-api-common
    #   vng-api-common
    #   zgw-consumers
django-axes==5.39.0
    # via -r requirements/base.txt
django-capture-on-commit-callbacks==1.11.0
    # via -r requirements/base.txt
django-choices==1.7.0
    # via
    #   -r requirements/base.txt
    #   django-simple-certmanager
    #   vng-api-common
    #   zgw-consumers
django-cors-middleware==1.5.0
    # via -r requirements/base.txt
django-filter==21.1
    # via
    #   -r requirements/base.txt
    #   vng-api-common
django-ipware==4.0.2
    # via
    #   -r requirements/base.txt
    #   django-axes
django-markup==1.3
    # via -r requirements/base.txt
django-privates==1.4.0
    # via
    #   -r requirements/base.txt
    #   django-simple-certmanager
django-redis==5.2.0
    # via -r requirements/base.txt
django-relativedelta==1.1.2
    # via
    #   -r requirements/base.txt
    #   zgw-consumers
django-rest-framework-condition==0.1.1
    # via
    #   -r requirements/base.txt
    #   vng-api-common
django-sendfile2==0.7.0
    # via
    #   -r requirements/base.txt
    #   django-privates
django-simple-certmanager==1.1.0
    # via
    #   -r requirements/base.txt
    #   zgw-consumers
django-solo==2.0.0
    # via
    #   -r requirements/base.txt
    #   notifications-api-common
    #   vng-api-common
    #   zgw-consumers
django-utils-six==2.0
    # via -r requirements/base.txt
djangorestframework==3.12.4
    # via
    #   -r requirements/base.txt
    #   drf-nested-routers
    #   drf-spectacular
    #   notifications-api-common
    #   vng-api-common
djangorestframework-camel-case==1.3.0
    # via
    #   -r requirements/base.txt
    #   notifications-api-common
    #   vng-api-common
docutils==0.17.1
    # via
    #   -r requirements/base.txt
    #   sphinx
    #   sphinx-rtd-theme
drf-nested-routers==0.93.4
    # via
    #   -r requirements/base.txt
    #   vng-api-common
drf-spectacular==0.24.2
    # via
    #   -r requirements/base.txt
    #   vng-api-common
faker==15.1.1
    # via
    #   -r requirements/base.txt
    #   zgw-consumers
gemma-zds-client==1.0.1
    # via
    #   -r requirements/base.txt
    #   notifications-api-common
    #   vng-api-common
    #   zgw-consumers
idna==3.4
    # via
    #   -r requirements/base.txt
    #   requests
imagesize==1.4.1
    # via
    #   -r requirements/base.txt
    #   sphinx
importlib-metadata==4.12.0
    # via
    #   -r requirements/base.txt
    #   markdown
    #   sphinx
inflection==0.5.1
    # via
    #   -r requirements/base.txt
    #   drf-spectacular
iso-639==0.4.5
    # via
    #   -r requirements/base.txt
    #   vng-api-common
isodate==0.6.1
    # via
    #   -r requirements/base.txt
    #   vng-api-common
jinja2==3.1.2
    # via
    #   -r requirements/base.txt
    #   sphinx
jsonschema==4.16.0
    # via
    #   -r requirements/base.txt
    #   drf-spectacular
markdown==3.4.1
    # via -r requirements/base.txt
markupsafe==2.1.1
    # via
    #   -r requirements/base.txt
    #   jinja2
notifications-api-common==0.1.0
    # via
    #   -r requirements/base.txt
    #   vng-api-common
oyaml==1.0
    # via
    #   -r requirements/base.txt
    #   vng-api-common
packaging==21.3
    # via
    #   -r requirements/base.txt
    #   redis
    #   sphinx
pillow==9.2.0
    # via -r requirements/base.txt
psycopg2-binary==2.9.3
    # via -r requirements/base.txt
pycparser==2.21
    # via
    #   -r requirements/base.txt
    #   cffi
pygments==2.13.0
    # via
    #   -r requirements/base.txt
    #   sphinx
pyjwt==2.5.0
    # via
    #   -r requirements/base.txt
    #   gemma-zds-client
    #   vng-api-common
pyopenssl==22.1.0
    # via
    #   -r requirements/base.txt
    #   django-simple-certmanager
    #   zgw-consumers
pyparsing==3.0.9
    # via
    #   -r requirements/base.txt
    #   packaging
pyrsistent==0.18.1
    # via
    #   -r requirements/base.txt
    #   jsonschema
python-dateutil==2.8.2
    # via
    #   -r requirements/base.txt
    #   django-relativedelta
    #   faker
python-dotenv==0.21.0
    # via -r requirements/base.txt
pytz==2022.2.1
    # via
    #   -r requirements/base.txt
    #   babel
    #   django
pyyaml==6.0
    # via
    #   -r requirements/base.txt
    #   drf-spectacular
    #   gemma-zds-client
    #   oyaml
    #   vng-api-common
raven==6.10.0
    # via -r requirements/base.txt
redis==4.3.4
    # via
    #   -r requirements/base.txt
    #   django-redis
requests==2.28.1
    # via
    #   -r requirements/base.txt
    #   gemma-zds-client
    #   requests-mock
    #   sphinx
    #   vng-api-common
    #   zgw-consumers
requests-mock==1.10.0
    # via
    #   -r requirements/base.txt
    #   zgw-consumers
six==1.16.0
    # via
    #   -r requirements/base.txt
    #   django-markup
    #   isodate
    #   python-dateutil
    #   requests-mock
snowballstemmer==2.2.0
    # via
    #   -r requirements/base.txt
    #   sphinx
sphinx==5.2.2
    # via
    #   -r requirements/base.txt
    #   sphinx-rtd-theme
sphinx-rtd-theme==1.0.0
    # via -r requirements/base.txt
sphinxcontrib-applehelp==1.0.2
    # via
    #   -r requirements/base.txt
    #   sphinx
sphinxcontrib-devhelp==1.0.2
    # via
    #   -r requirements/base.txt
    #   sphinx
sphinxcontrib-htmlhelp==2.0.0
    # via
    #   -r requirements/base.txt
    #   sphinx
sphinxcontrib-jsmath==1.0.1
    # via
    #   -r requirements/base.txt
    #   sphinx
sphinxcontrib-qthelp==1.0.3
    # via
    #   -r requirements/base.txt
    #   sphinx
sphinxcontrib-serializinghtml==1.1.5
    # via
    #   -r requirements/base.txt
    #   sphinx
sqlparse==0.4.3
    # via
    #   -r requirements/base.txt
    #   django
uritemplate==4.1.1
    # via
    #   -r requirements/base.txt
    #   drf-spectacular
urllib3==1.26.12
    # via
    #   -r requirements/base.txt
    #   requests
uwsgi==2.0.20
    # via -r requirements/production.in
vng-api-common==1.11.0
    # via -r requirements/base.txt
wrapt==1.14.1
    # via
    #   -r requirements/base.txt
    #   deprecated
zgw-consumers==0.21.2
    # via
    #   -r requirements/base.txt
    #   notifications-api-common
zipp==3.8.1
    # via
    #   -r requirements/base.txt
    #   importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
setuptools==65.4.1
    # via
    #   -r requirements/base.txt
    #   django-axes
